Added pokemon catch count, pokestop visit count and sniper snipe count to prevent being softbanned. Options added: Under "PokemonSettings": { "SnipeRestSeconds": 660, "SnipeCountLimit": 39, "CatchPokemonLimit": 999, "CatchPokemonLimitMinutes": 1470, "PokeStopLimit": 1998, "PokeStopLimitMinutes": 1470,

+++ PoGo.NecroBot.Logic/Tasks/FavoritePokemonTask.cs
@@ -29,7 +29,7 @@ namespace PoGo.NecroBot.Logic.Tasks
 
                 if (session.LogicSettings.AutoFavoritePokemon && perfection >= session.LogicSettings.FavoriteMinIvPercentage && pokemon.Favorite!=1)
                 {
-                    await session.Client.Inventory.SetFavoritePokemon(pokemon.Id, true);
+                    await session.Client.Inventory.SetFavoritePokemon((long)pokemon.Id, true);
 
                     session.EventDispatcher.Send(new NoticeEvent
                     {
